requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_py-xgboost-mutex=2.0=cpu_0
absl-py=1.4.0=pypi_0
anyio=3.6.2=pypi_0
appdirs=1.4.4=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
arviz=0.13.0=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
attrs=22.1.0=pyh71513ae_1
ax-platform=0.2.4=pyhd8ed1ab_0
babel=2.11.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
binaryornot=0.4.4=pypi_0
blas=1.0=mkl
bleach=5.0.1=pypi_0
boruta=0.3=pypi_0
boto3=1.26.120=pypi_0
botocore=1.29.120=pypi_0
botorch=0.6.2=pyhd8ed1ab_0
brotli=1.0.9=hcfcfb64_8
brotli-bin=1.0.9=hcfcfb64_8
bzip2=1.0.8=h8ffe710_4
ca-certificates=2022.12.7=h5b45459_0
cachetools=5.3.0=pypi_0
cdsapi=0.6.1=pypi_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py37ha95fbe2_1
cftime=1.6.2=py37h0b711f8_0
chardet=5.1.0=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.2.1=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
convertdate=2.4.0=pyhd8ed1ab_0
cookiecutter=2.1.1=pypi_0
curl=7.86.0=heaf79c2_1
cycler=0.11.0=pyhd8ed1ab_0
cython=3.0.0a11=pypi_0
dask=2022.2.0=pypi_0
debugpy=1.6.3=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.13=pyh6c4a22f_0
docker-pycreds=0.4.0=pypi_0
docutils=0.20.1=pypi_0
entrypoints=0.4=pypi_0
ephem=4.1.3=py37h51bd9d9_4
exceptiongroup=1.0.1=pypi_0
fastjsonschema=2.16.2=pypi_0
fbprophet=0.7.1=py37h1834ac0_0
flatbuffers=23.3.3=pypi_0
fonttools=4.38.0=py37h51bd9d9_0
freetype=2.12.1=h546665d_0
fsspec=2023.1.0=pypi_0
future=0.18.2=py37h03978a9_5
gast=0.4.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
google-auth=2.17.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
gpytorch=1.8.1=pyhd8ed1ab_0
greenlet=1.1.3=py37hf2a7229_0
grpcio=1.54.0=pypi_0
h5py=3.8.0=pypi_0
hdf4=4.2.15=h1b1b6ef_5
hdf5=1.12.2=nompi_h57737ce_100
hijri-converter=2.2.4=pyhd8ed1ab_0
holidays=0.16=pyhd8ed1ab_0
idna=3.4=pypi_0
imbalanced-learn=0.10.0=pyhd8ed1ab_0
importlib-metadata=4.11.4=py37h03978a9_0
importlib-resources=5.10.0=pypi_0
importlib_metadata=4.11.4=hd8ed1ab_0
iniconfig=1.1.1=pypi_0
intel-openmp=2022.1.0=h57928b3_3787
ipykernel=6.16.2=pypi_0
ipython=7.34.0=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=pyhd8ed1ab_1
jinja2-time=0.2.0=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h8ffe710_2
json5=0.9.10=pypi_0
jsonschema=4.17.0=pypi_0
jupyter-client=7.4.5=pypi_0
jupyter-core=4.11.2=pypi_0
jupyter-server=1.23.2=pypi_0
jupyterlab=3.5.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.16.3=pypi_0
kats=0.2.0=pyhd8ed1ab_0
keras=2.11.0=pypi_0
kiwisolver=1.4.4=py37h8c56517_0
korean_lunar_calendar=0.3.1=pyhd8ed1ab_0
krb5=1.19.3=hc8ab02b_0
lcms2=2.14=h90d422f_0
lerc=4.0.0=h63175ca_0
libblas=3.9.0=12_win64_mkl
libbrotlicommon=1.0.9=hcfcfb64_8
libbrotlidec=1.0.9=hcfcfb64_8
libbrotlienc=1.0.9=hcfcfb64_8
libcblas=3.9.0=12_win64_mkl
libclang=16.0.0=pypi_0
libcurl=7.86.0=heaf79c2_1
libdeflate=1.14=hcfcfb64_0
libiconv=1.17=h8ffe710_0
liblapack=3.9.0=12_win64_mkl
libnetcdf=4.8.1=nompi_h8c042bf_106
libpng=1.6.38=h19919ed_0
libpython=2.2=py37h03978a9_1
libsqlite=3.39.4=hcfcfb64_0
libssh2=1.10.0=h9a1e1f7_3
libtiff=4.4.0=h8e97e67_4
libuv=1.44.2=h8ffe710_0
libwebp-base=1.2.4=h8ffe710_0
libxcb=1.13=hcd874cb_1004
libxgboost=1.6.2=cpu_h20390bd_1
libxml2=2.10.3=hc3477c8_0
libzip=1.9.2=h519de47_1
libzlib=1.2.13=hcfcfb64_4
llvmlite=0.39.1=pypi_0
locket=1.0.0=pypi_0
lockfile=0.12.2=pypi_0
luigi=3.3.0=pypi_0
lunarcalendar=0.0.9=py_0
m2w64-binutils=2.25.1=5
m2w64-bzip2=1.0.6=6
m2w64-crt-git=5.0.0.4636.2595836=2
m2w64-gcc=5.3.0=6
m2w64-gcc-ada=5.3.0=6
m2w64-gcc-fortran=5.3.0=6
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gcc-objc=5.3.0=6
m2w64-gmp=6.1.0=2
m2w64-headers-git=5.0.0.4636.c0ad18a=2
m2w64-isl=0.16.1=2
m2w64-libiconv=1.14=6
m2w64-libmangle-git=5.0.0.4509.2e5a9a2=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
m2w64-make=4.1.2351.a80a8b8=2
m2w64-mpc=1.0.3=3
m2w64-mpfr=3.1.4=4
m2w64-pkg-config=0.29.1=2
m2w64-toolchain=5.3.0=7
m2w64-toolchain_win-64=2.4.0=0
m2w64-tools-git=5.0.0.4592.90b8472=2
m2w64-windows-default-manifest=6.4=3
m2w64-winpthreads-git=5.0.0.4634.697f757=2
m2w64-zlib=1.2.8=10
markdown=3.4.3=pypi_0
markupsafe=2.1.1=py37hcc03f2d_1
matplotlib-base=3.5.3=py37hbaab90a_2
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.4=pypi_0
mkl=2021.4.0=h0e2418a_729
mkl-service=2.4.0=py37h75fcce0_0
msys2-conda-epoch=20160418=1
multipledispatch=0.6.0=py_0
munkres=1.1.4=pyh9f0ad1d_0
nbclassic=0.4.8=pypi_0
nbclient=0.7.0=pypi_0
nbconvert=7.2.5=pypi_0
nbformat=5.7.0=pypi_0
nest-asyncio=1.5.6=pypi_0
netcdf4=1.6.1=nompi_py37he734059_100
ninja=1.11.0=h2d74725_0
notebook=6.5.2=pypi_0
notebook-shim=0.2.2=pypi_0
numba=0.56.3=py37h9c05cb2_0
numpy=1.21.6=py37h2830a78_0
oauthlib=3.2.2=pypi_0
opencv-python=4.7.0.72=pypi_0
openjpeg=2.5.0=hc9384bd_1
openssl=3.0.7=hcfcfb64_1
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.5=py37h9386db6_0
pandocfilters=1.5.0=pypi_0
parameterized=0.8.1=pyhd8ed1ab_1
parso=0.8.3=pypi_0
partd=1.4.0=pypi_0
pathtools=0.1.2=pypi_0
patsy=0.5.3=pyhd8ed1ab_0
pickleshare=0.7.5=pypi_0
pillow=9.3.0=pypi_0
pip=22.3.1=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pypi_0
plotly=5.11.0=pyhd8ed1ab_0
prometheus-client=0.15.0=pypi_0
prompt-toolkit=3.0.32=pypi_0
protobuf=3.19.6=pypi_0
psutil=5.9.4=pypi_0
psycopg2=2.9.6=pypi_0
pthread-stubs=0.4=hcd874cb_1001
py-xgboost=1.6.2=cpu_py37h48304b8_1
pyarrow=12.0.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pydot=1.4.2=pypi_0
pygments=2.13.0=pypi_0
pymannkendall=1.4.2=pyh8a188c0_0
pymeeus=0.5.11=pypi_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyrsistent=0.19.2=pypi_0
pystan=2.19.1.1=py37h631819c_3
python=3.7.12=h900ac77_100_cpython
python-daemon=3.0.1=pypi_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=0.21.1=pypi_0
python-graphviz=0.20.1=pypi_0
python-slugify=8.0.1=pypi_0
python_abi=3.7=2_cp37m
pytorch=1.10.2=cpu_py37h907fbb5_0
pytz=2022.6=pyhd8ed1ab_0
pywin32=305=pypi_0
pywinpty=2.0.9=pypi_0
pyyaml=6.0=pypi_0
pyzmq=24.0.1=pypi_0
quantile-forest=1.0.2=pypi_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
s3transfer=0.6.0=pypi_0
scikeras=0.10.0=pypi_0
scikit-learn=1.0.2=py37hcabfae0_0
scipy=1.7.3=py37hb6553fb_0
seaborn=0.12.1=hd8ed1ab_0
seaborn-base=0.12.1=pyhd8ed1ab_0
send2trash=1.8.0=pypi_0
sentry-sdk=1.21.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=67.8.0=pypi_0
setuptools-git=1.2=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.0=pypi_0
smogn=0.1.2=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlalchemy=1.4.42=py37h51bd9d9_0
sqlite=3.39.4=hcfcfb64_0
statsmodels=0.13.2=py37h3a130e4_0
tbb=2021.6.0=h91493d7_1
tenacity=8.1.0=pyhd8ed1ab_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.11.0=pypi_0
tensorflow-estimator=2.11.0=pypi_0
tensorflow-intel=2.11.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.3.0=pypi_0
terminado=0.17.0=pypi_0
text-unidecode=1.3=pypi_0
threadpoolctl=3.1.0=pyh8a188c0_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h8ffe710_0
tomli=2.0.1=pypi_0
toolz=0.12.0=pypi_0
torch=1.13.0=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.5.0=pypi_0
typeguard=2.13.3=pyhd8ed1ab_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
ucrt=10.0.22621.0=h57928b3_0
unicodedata2=14.0.0=py37hcc03f2d_1
urllib3=1.26.12=pypi_0
vc=14.3=h3d8a991_9
vs2015_runtime=14.32.31332=h1d6e394_9
wandb=0.15.0=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.4.2=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=pyhd8ed1ab_0
wincertstore=0.2=pyhd8ed1ab_1009
wrapt=1.14.1=py37hcc03f2d_0
xarray=0.20.2=pyhd8ed1ab_0
xarray-einstats=0.2.2=pyhd8ed1ab_0
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
zipp=3.10.0=pyhd8ed1ab_0
zlib=1.2.13=hcfcfb64_4
zstd=1.5.2=h7755175_4
